tts/tts-python-livekit/pyproject.toml
[project]
name = "tts-python-livekit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents[hume]>=1.1.0",
    "livekit-plugins-anthropic>=1.1.0",
    "livekit-plugins-groq>=1.1.0",
    "livekit-plugins-silero>=1.1.0",
    "python-dotenv>=1.1.0",
    "simpleaudio>=1.0.4",
]
